requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
qrcode==7.4.2
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyPDF2==3.0.1
reportlab==3.6.13
